requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.4.4
pymongo==4.5.0
psycopg2-binary==2.9.6
